+++ AngularClient/src/main/core/components/getting-started.component.ts
@@ -1,7 +1,6 @@
 import { Component, OnInit } from "@angular/core";
 
 import { Project } from "../entities/project";
-import { User } from "../entities/user";
 import { AuthService } from "../auth.service";
 import { ProjectService } from "../project.service";