+++ AngularClient/src/main/core/core.module.ts
@@ -4,7 +4,10 @@ import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
 import { RouterModule, Routes } from "@angular/router";
 import { Angulartics2Module } from "angulartics2";
 import { Angulartics2GoogleAnalytics } from "angulartics2/ga";
-import { BreezeBridgeAngularModule } from "../../libraries/breeze-bridge-angular";
+
+// Breeze
+import "./breeze-client-odata-fix";
+import { BreezeBridgeHttpClientModule } from "breeze-bridge2-angular";
 
 // Internal modules
 import { AppHttpClient, AppHttpClientModule } from "./app-http-client/app-http-client.module";
@@ -77,7 +80,7 @@ export function appInitializer(authService: AuthService, googleAnalyticsService:
     AppHttpClientModule,
     RouterModule.forRoot(coreRoutes),
     Angulartics2Module.forRoot([Angulartics2GoogleAnalytics]),
-    BreezeBridgeAngularModule
+    BreezeBridgeHttpClientModule
   ],
   providers: [
     // Application initializer