+++ AngularClient/src/main/core/entities/project.spec.ts
@@ -1,5 +1,3 @@
-import { TestHelpers } from "./test-helpers";
-
 describe("main/core/entities/project", () => {
 
 });