+++ BusinessObjects/Entities/ElementItem.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
+using System.Linq;
 
 namespace forCrowd.Backbone.BusinessObjects.Entities
 {
@@ -26,5 +27,22 @@ namespace forCrowd.Backbone.BusinessObjects.Entities
         public ICollection<ElementCell> ParentCellSet { get; set; }
 
         string name;
+
+        public ElementCell AddCell(ElementField field)
+        {
+            Framework.Validations.ArgumentNullOrDefault(field, nameof(field));
+
+            if (ElementCellSet.Any(item => item.ElementField == field))
+                throw new System.Exception("An element item can't have more than one cell for the same field.");
+
+            var cell = new ElementCell
+            {
+                ElementField = field,
+                ElementItem = this
+            };
+            field.ElementCellSet.Add(cell);
+            ElementCellSet.Add(cell);
+            return cell;
+        }
     }
 }